Move player related entities to separate package
