+++ ippon/permissions_test.py
@@ -9,64 +9,10 @@ from rest_framework.test import APIClient
 
 import ippon.club.permissisons
 import ippon.permissions as permissions
-import ippon.serializers as serializers
-from ippon.models import Team, Player, TeamFight, TournamentAdmin, Tournament
+from ippon.models import Team, TeamFight, TournamentAdmin, Tournament
+import ippon.player.models as plm
 import ippon.club.models as cl
 from ippon.serializers import PointSerializer
-
-
-class PlayerPermissionsTest(django.test.TestCase):
-    def setUp(self):
-        self.user = User.objects.create(username='user', password='password')
-        self.admin = User.objects.create(username='admin', password='password')
-        self.club = cl.Club.objects.create(
-            name='cn1',
-            webpage='http://cw1.co',
-            description='cd1',
-            city='cc1')
-        self.player = self.club.players.create(name='pn1', surname='ps1', rank=7,
-                                               birthday=datetime.date(year=2001, month=1, day=1), sex=1,
-                                               club_id=self.club)
-        self.permission = ippon.club.permissisons.IsClubAdminOrReadOnlyDependent()
-        self.request = unittest.mock.Mock(user=self.user, data=serializers.PlayerSerializer(self.player).data)
-        self.view = unittest.mock.Mock()
-        self.club_admin = self.club.admins.create(user=self.admin)
-
-
-class PlayerPermissionTestsNotAdmin(PlayerPermissionsTest):
-    def setUp(self):
-        super(PlayerPermissionTestsNotAdmin, self).setUp()
-
-    def test_permits_when_safe_method(self):
-        self.request.method = 'GET'
-        result = self.permission.has_object_permission(self.request, self.view, self.player)
-        self.assertEqual(result, True)
-
-    def test_doesnt_permit_when_unsafe_method(self):
-        self.request.method = 'PUT'
-        result = self.permission.has_object_permission(self.request, self.view, self.player)
-        self.assertEqual(result, False)
-
-    def test_doesnt_permit_when_post(self):
-        self.request.method = 'POST'
-        result = self.permission.has_permission(self.request, self.view)
-        self.assertEqual(result, False)
-
-
-class PlayerPermissionTestsAdmin(PlayerPermissionsTest):
-    def setUp(self):
-        super(PlayerPermissionTestsAdmin, self).setUp()
-        self.club.admins.create(user=self.user)
-
-    def test_permits_when_safe_method(self):
-        self.request.method = 'GET'
-        result = self.permission.has_object_permission(self.request, self.view, self.player)
-        self.assertEqual(result, True)
-
-    def test_permits_when_unsafe_method(self):
-        self.request.method = 'PUT'
-        result = self.permission.has_object_permission(self.request, self.view, self.player)
-        self.assertEqual(result, True)
 
 
 class TournamentPermissionTests(django.test.TestCase):
@@ -565,11 +511,11 @@ class TestPointPermissions(django.test.TestCase):
                                             age_constraint_value=20, rank_constraint=5, rank_constraint_value=7,
                                             sex_constraint=1)
         self.t1 = Team.objects.create(tournament=self.to, name='t1')
-        self.p1 = Player.objects.create(name='pn1', surname='ps1', rank=7,
+        self.p1 = plm.Player.objects.create(name='pn1', surname='ps1', rank=7,
                                         birthday=datetime.date(year=2001, month=1, day=1), sex=1, club_id=c)
         self.t1.team_members.create(player=self.p1)
         self.t2 = Team.objects.create(tournament=self.to, name='t2')
-        self.p2 = Player.objects.create(name='pn2', surname='ps2', rank=7,
+        self.p2 = plm.Player.objects.create(name='pn2', surname='ps2', rank=7,
                                         birthday=datetime.date(year=2001, month=1, day=1), sex=1, club_id=c)
         self.t2.team_members.create(player=self.p2)
