+++ ippon/player/shallow_player_views_test.py
@@ -6,7 +6,7 @@ from django.urls import reverse
 from rest_framework import status
 from rest_framework.test import APIClient, APITestCase
 
-from ippon.models import Player
+import ippon.player.models as plm
 import ippon.club.models as cl
 
 BAD_PK = 0
@@ -17,12 +17,12 @@ class ShallowPlayerViewTest(APITestCase):
         self.client = APIClient()
         self.u1 = User.objects.create(username='a1', password='password1')
         self.c1 = cl.Club.objects.create(name='cn1', webpage='http://cw1.co', description='cd1', city='cc1')
-        self.p1 = Player.objects.create(name='pn1', surname='ps1', rank=1,
-                                        birthday=datetime.date(year=2001, month=1, day=1), sex=1, club_id=self.c1)
-        self.p2 = Player.objects.create(name='pn2', surname='ps2', rank=2,
-                                        birthday=datetime.date(year=2002, month=2, day=2), sex=0, club_id=self.c1)
-        self.p3 = Player.objects.create(name='pn3', surname='ps3', rank=3,
-                                        birthday=datetime.date(year=2003, month=3, day=3), sex=1, club_id=self.c1)
+        self.p1 = plm.Player.objects.create(name='pn1', surname='ps1', rank=1,
+                                            birthday=datetime.date(year=2001, month=1, day=1), sex=1, club_id=self.c1)
+        self.p2 = plm.Player.objects.create(name='pn2', surname='ps2', rank=2,
+                                            birthday=datetime.date(year=2002, month=2, day=2), sex=0, club_id=self.c1)
+        self.p3 = plm.Player.objects.create(name='pn3', surname='ps3', rank=3,
+                                            birthday=datetime.date(year=2003, month=3, day=3), sex=1, club_id=self.c1)
         self.valid_payload = {
             "name": "pn7",
             "surname": "ps7",
