+++ ippon/teams_views_test.py
@@ -6,7 +6,8 @@ from django.urls import reverse
 from rest_framework import status
 from rest_framework.test import APIClient, APITestCase
 
-from ippon.models import Player, Team, TournamentAdmin, Tournament
+from ippon.models import Team, TournamentAdmin, Tournament
+import ippon.player.models as plm
 import ippon.club.models as cl
 
 BAD_PK = 0
@@ -24,17 +25,17 @@ class TeamsViewTest(APITestCase):
                                             sex_constraint=1)
         self.t1 = Team.objects.create(tournament=self.to, name='t1')
         self.c = cl.Club.objects.create(name='cn1', webpage='http://cw1.co', description='cd1', city='cc1')
-        self.p1 = Player.objects.create(name='pn1', surname='ps1', rank=7,
-                                        birthday=datetime.date(year=2001, month=1, day=1), sex=1, club_id=self.c)
+        self.p1 = plm.Player.objects.create(name='pn1', surname='ps1', rank=7,
+                                            birthday=datetime.date(year=2001, month=1, day=1), sex=1, club_id=self.c)
         self.to.participations.create(player=self.p1)
         self.t1.team_members.create(player=self.p1)
         self.t2 = Team.objects.create(tournament=self.to, name='t2')
-        self.p2 = Player.objects.create(name='pn2', surname='ps2', rank=7,
-                                        birthday=datetime.date(year=2001, month=1, day=1), sex=1, club_id=self.c)
+        self.p2 = plm.Player.objects.create(name='pn2', surname='ps2', rank=7,
+                                            birthday=datetime.date(year=2001, month=1, day=1), sex=1, club_id=self.c)
         self.to.participations.create(player=self.p2)
         self.t2.team_members.create(player=self.p2)
-        self.p3 = Player.objects.create(name='pn3', surname='ps3', rank=7,
-                                        birthday=datetime.date(year=2001, month=1, day=1), sex=1, club_id=self.c)
+        self.p3 = plm.Player.objects.create(name='pn3', surname='ps3', rank=7,
+                                            birthday=datetime.date(year=2001, month=1, day=1), sex=1, club_id=self.c)
         self.to.participations.create(player=self.p3)
         self.t2.team_members.create(player=self.p3)
         self.t1_json = {'id': self.t1.id, 'tournament': self.to.id, 'members': [self.p1.id], 'name': 't1'}
@@ -93,8 +94,8 @@ class TeamViewSetAuthorizedTests(TeamsViewTest):
         self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
 
     def test_teams_get_non_members_returns_list_of_participants_not_assigned_to_a_team(self):
-        p4 = Player.objects.create(name='pn4', surname='ps4', rank=7,
-                                   birthday=datetime.date(year=2004, month=4, day=4), sex=1, club_id=self.c)
+        p4 = plm.Player.objects.create(name='pn4', surname='ps4', rank=7,
+                                       birthday=datetime.date(year=2004, month=4, day=4), sex=1, club_id=self.c)
         self.to.participations.create(player=p4)
 
         response = self.client.get(reverse('team-not-assigned', kwargs={'pk': self.t1.pk}))
@@ -191,8 +192,8 @@ class TeamViewSetUnauthenticatedTests(TeamsViewTest):
         self.assertEqual(response.status_code, status.HTTP_200_OK)
 
     def test_teams_get_non_members_returns_unauthorized(self):
-        p4 = Player.objects.create(name='pn4', surname='ps4', rank=7,
-                                   birthday=datetime.date(year=2004, month=4, day=4), sex=1, club_id=self.c)
+        p4 = plm.Player.objects.create(name='pn4', surname='ps4', rank=7,
+                                       birthday=datetime.date(year=2004, month=4, day=4), sex=1, club_id=self.c)
         self.to.participations.create(player=p4)
 
         response = self.client.get(reverse('team-not-assigned', kwargs={'pk': self.t1.pk}))
@@ -223,8 +224,8 @@ class ValidIdsTeamMembersViewTests(AuthorizedTeamMembersViewTests):
     def setUp(self):
         super(ValidIdsTeamMembersViewTests, self).setUp()
         self.t1 = Team.objects.create(tournament=self.to, name='t1')
-        self.p1 = Player.objects.create(name='pnn1', surname='pss1', rank=7,
-                                        birthday=datetime.date(year=2001, month=1, day=1), sex=1, club_id=self.club)
+        self.p1 = plm.Player.objects.create(name='pnn1', surname='pss1', rank=7,
+                                            birthday=datetime.date(year=2001, month=1, day=1), sex=1, club_id=self.club)
 
     def test_post_valid_payload_creates_specified_teammember(self):
         response = self.client.post(
@@ -245,8 +246,8 @@ class InvalidIdsTeamMemberViewTests(AuthorizedTeamMembersViewTests):
         super(InvalidIdsTeamMemberViewTests, self).setUp()
 
     def test_teams_post_invalid_team_returns_404(self):
-        p1 = Player.objects.create(name='pn1', surname='ps1', rank=7,
-                                   birthday=datetime.date(year=2001, month=1, day=1), sex=1, club_id=self.club)
+        p1 = plm.Player.objects.create(name='pn1', surname='ps1', rank=7,
+                                       birthday=datetime.date(year=2001, month=1, day=1), sex=1, club_id=self.club)
 
         response = self.client.post(
             reverse('team-members', kwargs={'pk': BAD_PK, 'player_id': p1.id}),
@@ -263,7 +264,7 @@ class InvalidIdsTeamMemberViewTests(AuthorizedTeamMembersViewTests):
         self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
 
     def test_delete_not_existing_team_member_returns_bad_request(self):
-        p1 = Player.objects.create(name='pn1', surname='ps1', rank=7,
-                                   birthday=datetime.date(year=2001, month=1, day=1), sex=1, club_id=self.club)
+        p1 = plm.Player.objects.create(name='pn1', surname='ps1', rank=7,
+                                       birthday=datetime.date(year=2001, month=1, day=1), sex=1, club_id=self.club)
         response = self.client.delete(reverse('team-members', kwargs={'pk': BAD_PK, 'player_id': p1.pk}))
         self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
